Hide state attribute to allow sane handling from API
This commit hides the "state" attribute of partial results
and implements a Setter for it.
This way it can remember if the state was set explicitely
or not.

In combination with the default state, it is now possible
to set a default state which is used, if NO partial results and NO
explicit states are set later.

An example:

var subcheck PartialResult
subcheck.SetDefaultState(check.Warning)

will create a PartialResult which returns Warning if nothing else
happens.
If I now, later in the code either add PartialResults or set the state
explicitely, this is overriden.

This way application code does not need to remember in what state
it put a PartialResult previously, but can just set a sane default
at the beginning.