Add initializer for PartialResult
This commit adds a NewPartialResult function which returns a
PartialResult struct with "sane" default, mainly setting the
defaultState to "Unknown".

This should help developers implementing this library, since it
avoids the following pattern:

sc := result.PartialResult{}
sc.setDefaultState(check.Unknown)

This pattern is quite common, but in a sense useless.